TextBox develop.

See merge request crupest/CruUI!2

+++ CruUI/base.cpp
@@ -13,7 +13,7 @@ namespace cru
         const auto length = ::WideCharToMultiByte(CP_UTF8, 0, string.data(), -1, nullptr, 0, nullptr, nullptr);
         MultiByteString result;
         result.reserve(length);
-        if (::WideCharToMultiByte(CP_UTF8, 0, string.data(), -1, result.data(), result.capacity(), nullptr, nullptr) == 0)
+        if (::WideCharToMultiByte(CP_UTF8, 0, string.data(), -1, result.data(), static_cast<int>(result.capacity()), nullptr, nullptr) == 0)
             throw Win32Error(::GetLastError(), "Failed to convert wide string to UTF-8.");
         return result;
     }
